move player respawn shader control to PlayerSpawnManager
