+++ Assets/Scripts/Player/PlayerSpawnManager.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Collections;
 
 using UnityEngine;
 
@@ -15,6 +16,9 @@ namespace Player
         private Spawn _currentSpawnPoint;
         public Room CurrentRoom => _currentRoom;
 
+        private SpriteRenderer _spriteR;
+        [SerializeField] private float spawnAnimationTime = .5f;
+
         public event Action OnPlayerRespawn;
 
         public Spawn CurrentSpawnPoint
@@ -29,14 +33,20 @@ namespace Player
             }
         }
 
+        void Start() {
+            _spriteR = GetComponentInChildren<SpriteRenderer>();
+        }
+
         private void OnEnable()
         {
             Room.RoomTransitionEvent += OnRoomTransition;
+            OnPlayerRespawn += ShaderRespawn;
         }
 
         private void OnDisable()
         {
             Room.RoomTransitionEvent -= OnRoomTransition;
+            OnPlayerRespawn -= ShaderRespawn;
         }
 
         public void Respawn()
@@ -56,6 +66,21 @@ namespace Player
             _prevRoom = _currentRoom;
             _currentRoom = roomEntering;
             _currentSpawnPoint = FindClosestSpawnPoint();
+        }
+
+        private IEnumerator ShaderRespawnCo() {
+            _spriteR.material.SetFloat("_Progress", 0);
+            float timer = 0;
+            while (timer < spawnAnimationTime) {
+                timer += Time.deltaTime;
+                _spriteR.material.SetFloat("_Progress", timer/spawnAnimationTime);
+                yield return null;
+            }
+            _spriteR.material.SetFloat("_Progress", 2);
+        }
+
+        public void ShaderRespawn() {
+            StartCoroutine(ShaderRespawnCo());
         }
 
         private Spawn FindClosestSpawnPoint()
